# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Breathe ESG – data ingestion and analyst review platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  slug: string;
  createdAt: string;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}

export type IngestionBatchSourceType = typeof IngestionBatchSourceType[keyof typeof IngestionBatchSourceType];


export const IngestionBatchSourceType = {
  sap: 'sap',
  utility: 'utility',
  travel: 'travel',
} as const;

export type IngestionBatchStatus = typeof IngestionBatchStatus[keyof typeof IngestionBatchStatus];


export const IngestionBatchStatus = {
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface IngestionBatch {
  id: number;
  clientId: number;
  sourceType: IngestionBatchSourceType;
  status: IngestionBatchStatus;
  filename: string;
  rowCount: number;
  errorCount: number;
  /** @nullable */
  errorDetails?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export type EmissionsRecordScope = typeof EmissionsRecordScope[keyof typeof EmissionsRecordScope];


export const EmissionsRecordScope = {
  scope1: 'scope1',
  scope2: 'scope2',
  scope3: 'scope3',
} as const;

export type EmissionsRecordSourceType = typeof EmissionsRecordSourceType[keyof typeof EmissionsRecordSourceType];


export const EmissionsRecordSourceType = {
  sap: 'sap',
  utility: 'utility',
  travel: 'travel',
} as const;

export type EmissionsRecordRawData = { [key: string]: unknown };

export type EmissionsRecordStatus = typeof EmissionsRecordStatus[keyof typeof EmissionsRecordStatus];


export const EmissionsRecordStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  flagged: 'flagged',
} as const;

export interface EmissionsRecord {
  id: number;
  batchId: number;
  clientId: number;
  scope: EmissionsRecordScope;
  category: string;
  sourceType: EmissionsRecordSourceType;
  rawData?: EmissionsRecordRawData;
  activityAmount: number;
  activityUnit: string;
  normalizedCo2eKg: number;
  activityDate: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  status: EmissionsRecordStatus;
  /** @nullable */
  reviewNotes?: string | null;
  /** @nullable */
  reviewedBy?: string | null;
  /** @nullable */
  reviewedAt?: string | null;
  isSuspicious: boolean;
  /** @nullable */
  suspiciousReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RecordList {
  data: EmissionsRecord[];
  total: number;
  page: number;
  limit: number;
}

export type RecordReviewInputStatus = typeof RecordReviewInputStatus[keyof typeof RecordReviewInputStatus];


export const RecordReviewInputStatus = {
  approved: 'approved',
  rejected: 'rejected',
  flagged: 'flagged',
} as const;

export interface RecordReviewInput {
  status: RecordReviewInputStatus;
  reviewNotes?: string;
  reviewedBy?: string;
}

export type AuditLogEntryAction = typeof AuditLogEntryAction[keyof typeof AuditLogEntryAction];


export const AuditLogEntryAction = {
  approved: 'approved',
  rejected: 'rejected',
  flagged: 'flagged',
  edited: 'edited',
  deleted: 'deleted',
} as const;

export type AuditLogEntryBeforeState = { [key: string]: unknown };

export type AuditLogEntryAfterState = { [key: string]: unknown };

export interface AuditLogEntry {
  id: number;
  recordId: number;
  action: AuditLogEntryAction;
  actor: string;
  beforeState?: AuditLogEntryBeforeState;
  afterState?: AuditLogEntryAfterState;
  createdAt: string;
}

export type ScopeBreakdownScope = typeof ScopeBreakdownScope[keyof typeof ScopeBreakdownScope];


export const ScopeBreakdownScope = {
  scope1: 'scope1',
  scope2: 'scope2',
  scope3: 'scope3',
} as const;

export interface ScopeBreakdown {
  scope: ScopeBreakdownScope;
  count: number;
  co2eKg: number;
}

export type SourceBreakdownSourceType = typeof SourceBreakdownSourceType[keyof typeof SourceBreakdownSourceType];


export const SourceBreakdownSourceType = {
  sap: 'sap',
  utility: 'utility',
  travel: 'travel',
} as const;

export interface SourceBreakdown {
  sourceType: SourceBreakdownSourceType;
  count: number;
  co2eKg: number;
}

export interface DashboardSummary {
  totalRecords: number;
  pendingCount: number;
  approvedCount: number;
  rejectedCount: number;
  flaggedCount: number;
  suspiciousCount: number;
  totalCo2eKg: number;
  byScope: ScopeBreakdown[];
  bySource: SourceBreakdown[];
  recentBatches: IngestionBatch[];
}

export interface SapIngestInput {
  clientId: number;
}

export interface UtilityIngestInput {
  clientId: number;
}

export interface TravelIngestInput {
  clientId: number;
}

export type ListBatchesParams = {
clientId?: number;
sourceType?: ListBatchesSourceType;
status?: ListBatchesStatus;
};

export type ListBatchesSourceType = typeof ListBatchesSourceType[keyof typeof ListBatchesSourceType];


export const ListBatchesSourceType = {
  sap: 'sap',
  utility: 'utility',
  travel: 'travel',
} as const;

export type ListBatchesStatus = typeof ListBatchesStatus[keyof typeof ListBatchesStatus];


export const ListBatchesStatus = {
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type ListRecordsParams = {
clientId?: number;
batchId?: number;
status?: ListRecordsStatus;
scope?: ListRecordsScope;
sourceType?: ListRecordsSourceType;
isSuspicious?: boolean;
page?: number;
limit?: number;
};

export type ListRecordsStatus = typeof ListRecordsStatus[keyof typeof ListRecordsStatus];


export const ListRecordsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  flagged: 'flagged',
} as const;

export type ListRecordsScope = typeof ListRecordsScope[keyof typeof ListRecordsScope];


export const ListRecordsScope = {
  scope1: 'scope1',
  scope2: 'scope2',
  scope3: 'scope3',
} as const;

export type ListRecordsSourceType = typeof ListRecordsSourceType[keyof typeof ListRecordsSourceType];


export const ListRecordsSourceType = {
  sap: 'sap',
  utility: 'utility',
  travel: 'travel',
} as const;

export type GetDashboardSummaryParams = {
clientId?: number;
};