# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Breathe ESG – data ingestion and analyst review platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a single client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List ingestion batches
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: sourceType
          in: query
          schema:
            type: string
            enum: [sap, utility, travel]
        - name: status
          in: query
          schema:
            type: string
            enum: [processing, completed, failed]
      responses:
        "200":
          description: List of batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IngestionBatch"
  /batches/{id}:
    get:
      operationId: getBatch
      tags: [batches]
      summary: Get a single ingestion batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionBatch"
        "404":
          description: Not found
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
    ClientInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
    IngestionBatch:
      type: object
      required: [id, clientId, sourceType, status, filename, rowCount, errorCount, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        sourceType:
          type: string
          enum: [sap, utility, travel]
        status:
          type: string
          enum: [processing, completed, failed]
        filename:
          type: string
        rowCount:
          type: integer
        errorCount:
          type: integer
        errorDetails:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
...